feat(scripts: debug): add a script to start a debug session
Since switching to a higher-half kernel, placing the "target remote ..."
inside the gdbinit script seems to make GDB crash. Having it executed
as a CLI argument seems to be a working alternative.

This script lets us chose which symbol we want to set a breakpoint on,
instead of hardcoding it to kernel_main inside the gdbinit script.