feat(kernel: symbols): embed kernel symbol table
We now have a way to retrieve a symbol given its address!

We include the kernel's symbol table inside a dedicated section after
compiling the kernel.

This can be used to compute a meaningful stacktrace, but also to be able
to compile a program using the kernel's symbols (maybe later, idk if it
should really used like that ...).

Inspiration:
* https://github.com/SerenityOS/serenity/blob/master/Meta/gn/secondary/Kernel/post_process_kernel.py