feat(libc: printf): add support for '#'
man 3 printf:

Flag Characters : '#'

The  value  should  be converted to an "alternate form".  For o conversions,
the first character of the output string is made zero (by prefixing a 0 if it was
not zero already).  For x and X conversions, a nonzero result has the string "0x"
(or "0X" for X conversions) prepended to it.