feat(lib: libalgo): add AVL tree implementation
I plan to use an AVL to store VMAs inside the VMM.

This implementation is not responsible for allocating the node's
structure, as it is intended to be used with intrusive C structs.
The caller has to pass the address of the field of a bigger structure,
reprensenting the node's data.

This is inpired by linux's intrusive linked list design:
* https://www.data-structures-in-practice.com/intrusive-linked-lists

I used this opportunity to create a separate "libalgo" library, which
will contain the implementation of all the datastructures we need and
use throughout this project.

Signed-off by: Léo DUBOIN <leo@duboin.com>