feat(kernel: memory): implement VMM
This commit marks the first version of our vitual memory manager.

This implementation uses 2 AVL trees to keep track of the allocated
virtual areas. As this kernel is still young, and I'm still learning,
the actual content of the VMA structure will surely change a lot.

For now, we create an allocated region fo reach succesfull call to the
api, and merge the adjacent free regions whenever possible when freeing.

This has the advantage of limiting memory fragmentation, but causes a
lot of modification to be made on the trees, which can end up being costly.
As I said, the design is subject to change once I learn more but, for now,
we finally have a way to perform dynamic allocation inside our kernel!

Signed-off by: Léo DUBOIN <leo@duboin.com>