feat(memory: vmm: allocate): add optional starting address
We added an optional argument for the caller to specify a minimum
starting address for the returned area. If an address was explicetly
specified, the returned virtual area's starting address MUST be
superior or equal to it.

This change was added to be able to conform to mmap's POSIX definition.