fix(kernel: linker): reserve 32K of space for our symbols
The address of the _kernel_code_end is not retroactively udpated
when we fill the .kernel_symbols section using objcopy. This means
that if the kernel's symbol map grows too big, its end is not included
inside the identity map performed when initializing paging.

This caused a #PF when trying to read information about those symbols.

For now, we reserved 32K of space, which should be enough for a while.
We should either find a solution to relocate the _kernel_code_end
symbol, or think of another way to pass the symbol map.