feat(kernel: kmalloc): check against multi-freeing corruption
By freeing a valid address multiple times, a user could corrupt the
number of allocated areas inside a bucket. This could lead to the
bucket been unmapped, causing a #PF to be raised whenever a valid
block inside the now unmapped bucket is referenced.

To avoid that, I added an arbitrary magic number following the intrusive
linked list's node placed inside a free block. We check for the presence
of this value before checking, and prevent further actions if it is
found.