feat(kernel: mmu): stop hardcoding page directory
This is a necessary first step for swithing into a new process.

We now use a static page directory to startup our kernel, used for
setting up our first kernel pages. When creating a new page directory,
we then copy the 0xC000-0xFFFF range of page tables into the new one.

Once paging has been activated, we refer to the current page directory
by using the recursive mapping trick, so we don't even need to keep
track of the current PD's address in a context.

And ... this lets us remove our first batch of FIXMEs :^)