feat(kernel: vmm): split process VMM and kernel VMM
To allow sharing kernel pages, it is now necessary to also share a
common virtual memory manager used for allocating them.

We also reserve another KERNEL_VMM_RESERVED range inside the
virtual kernel address space. This range is used by the kernel vmm
instead of the usual "0x0010_0000:0x0011_0000" for allocating VMAs.
This is necessary, since the VMAs also need to be shared across
processes.

This was the occasion to remove the fixmes about "do not hardcode the
vmm".

TODO:
* This process should be generalised to: MAP_SHARED once we implement it