feat(kernel: fs): create virtual filesystem
At this stage it is only possible to mount a filesystem and find one of
its components. But in the future it should be possible to perform more
actions on the virtual nodes (such as open/close for example). This is
only to setup the basis for us to increment on later (as most of these
features are not needed, nor implementable as is).

This implementation comes with a POC in the form of a TAR filesystem.
It has the advantage of being simple to implement and to generate assets
for.

This implementation relies on a stub `dev_t` structure, since we only
need to write/read from RAM. It will be modified later to be able to
use different kind of devices as support (or at least one proper one).