feat(kernel: acpi): implement uACPI's kernel API
Some functions are still missing and marked as UNIMPLEMENTED.
Most of these are marked as such because we are missing the key
components of our kernel required to implement them.

These functions display a message onto the console when called
to be able to debug any eventual problem this may cause.