refacto(kernel: interrupts): allow generic IRQ handlers
We can now pass a generic (void *) argument to the handler when defining
it. If this parameter is omitted, we fallback to the old method of
calling the handler with the frame pointer to not break our previous
implementation.

It is now also possible to set a handler for all 255 entries of the IDT.
This means that the VTable is that much larger, as well as the number of
IRQ wrapper, but it should be good enough for now ig.