refacto(kernel: vfs): use address range instead of device
The design which used a device for loading a filesystem
does not correspond match with current device API design.

To not over engineer things (too late), and still be able to
use our filesystem as is, we switched to a [start:end] range
design instead, until we come up with something better.