fix(kernel: x86: mmu): pre-allocate shared page tables
Since we did not pre-allocate all of them, newly created page directory
entries were not propagated to previously created processes.

This notably caused a pagefault when an interrupt was triggered during
the IDLE task's time slice, since it is created very early (before most
of the page tables were allocated).