fix(kernel: mmu): offset map the first 1MiB
We previously identity mapped the first 1MiB of our address
space because it contains physical addresses linked to hardware
devices (VGA buffer).

Problem is this first 1MiB of address space was not duplicated in
each newly created page directory after that. Now, instead of going
through the hassle of duplicating it, we offset_map it alongside the
kernel's code during startup, and modified all references to it to
now use the offseted address instead.

Test: Use the VGA buffer without crashing, and test if VMM structures
      (that now use the first MiB) can be allocated correctly.