feat(kernel: file): add file structure
The struct file represents an opened file.

It will hold a high importance in the future, as it will be used
not only for user requested files, but also for interacting with
the device's content. According to the Unix philosophy, everything
is a file. This holds true for hardware devices, that fit the
"open -> read/write -> close" pattern perfectly.

At least that's how I envision things as of writing this, we'll see
if it holds true later.