refacto(kernel: log): use file-wide domain macro definition
Instead of passing in the domain name for each call to the log API
we now define a single LOG_DOMAIN macro at the top of a translation
unit. This macro is automatically inserted into each log API calls,
if present.

We also took this opportunity to stop always expanding the level macros
into their respective ANSI code + prefix, and instead use an enum which
is in term translated into the same representation. This reduces the
size of the final binary.