feat(kernel): add ELF32 executable loader
The Executable and Linkable Format is the most used
executable format on *NIX systems.

This commit does not add full support for this file format yet,
as we still need to implement dynamic linking (shared library),
and relocatable files (modules).

Nevertheless, it is now possible to load an ELF32 executable's content
in memory and jump onto its entrypoint, effectively executing it.

NOTE: We currently do not unload the ELF's content once done. This will
      surely cause an issue when forking. But it works for now.