feat(kernel: i686: process): always allocate a user stack
This is a temporary workaround allowing us to test the execution
of userland executables from a kernel process.

As marked inside the newly added comment, we should not be able to
create kernel processes anyway. This design is just a product of my
misunderstanding of process/threads + needing a quick first process API.

This will eventually be refactored when correctly implementing threads.