chore(kernel: process): rename process API into threads
What we are currently interacting is closer to a thread than a true
process. Renaming all references, structures and functions to fit
this view will make adding a process API easier.

Now we 'simply' have to extract the components of a process from the
thread structure instead of modifying everything else to fit the new
API.