feat(kernel: i686): increase the size of the bootstrap page tables
We're going to increase the total size of the kernel by quite a lot
when switching from a pageframe bitmap over to an array of struct page.

The current bootstrap model will not fit the entirety of our kernel,
causing us to overwrite some of the kernel's content when memsetting
the page array to 0.

This commit increases the maximum relocatable size, by mapping
4 consecutive page directory entries at a time instead of 1. This means
that the first 16MiB of physical memory is repeated across the entirety
of the virtual address space.