feat(kernel: pmm): add struct page
In the future, we will require to hold some information about
the physical pageframes (other than whether it is currently allocated).
We have thus replaced the bitmap, previously used to store
this simple information, by an array of page structs.

This array is statically allocated, and its size depends on the
theoretical maximum number of available pageframes. Compared to the
previous design, this one uses a lot more memory to store all this
information (+1MB).