fix(i686: process): invalid kernel stack address
We were storing the kernel stack's bottom address inside the process
context structure. This meant that it was also loaded inside the
process's TSS.ESP0, and used as the top of the kernel's stack when
switching from userland to kernel space.

This caused the metadata used by malloc to be corrupted when eventually
pushing values onto the kernel's stack.

It was the reason why our kernel kept randomly crashing when switching
between kernel/userland proceses !

Fixes 71259243c929b