fix(kernel: vmm): extract VMA with fixed address instead of splitting
When allocating a VMA using a fixed starting address, if the newly
allocated VMA was located inside a larger VMA, we did not reinsert
the part of the original VMA that was located after the now allocated
section.

This commit modifies the vmm_allocate function to re-use the helper
already used by vmm_free, that does exactly that already.