feat(memory): add address_space API
This API defines a process's address space.

Instead of using the VMM directly, each process (including the kernel
process) comes with a full-on address space abstraction. This address
space can be created, modified destroyed completely, copied or cloned
(see address_space_*() functions).

To allocate memory inside a given address space, we must now use the
vm_*() functions. They completely abstract away the previously needed,
and often done together, calls to the PMM+VMM+MMU API (simiar to
kmalloc, but on a page-wide level).

This commit contains a lot of different things that I was to lazy to
split into separate commts:
- memory/vm: Add address space (AS) API
- vmm, mmu: modify API to fit the new address_space (AS) api
- kmalloc: modify API to use the new vm_*() functions
- kmalloc: replace kmalloc_at()/kfree_at() with vm_alloc_at()/vm_free()
- process: initialize its address space when creating a new process

NOTE: As we are targetting a 32b platform, we do not have a quick way of
modifying the content located at an arbitrary physical address inside memory.
This means that to modify the content of another address space's virtual memory
space, it is required to allocate a new virtual address and remap the
original physical address. This is really slow, and should be changed.
We could for example reserved a bunch of virtual addresses for quickly
mapping short-lived pages.