fix(kernel: vmm: mmap_file): align length argument onto page size
This parameter should be checked and we should return an error, just
as regular mmap does, but all our file mapping API is kind of a
placeholder for now.

Since we do not make sure that files are sized in blocks, nor do we
have a proper mmap file_ops function, we cannot map files whose size
is not aligned onto a page size.

This will do for now, but should be removed.