feat(vfs): add socket pseudo-filesystem
By design, sockets need to be accessed just like any other regular
file. This meant that weneed to tallocate vnodes for them as well
as define file operations. This is exactly the purpose of this
pseudo-filesystem.

This pseudo-filesystem is not meant to be exposed, mounted, etc
(at least for now). It is only meant to allocate standalone vnodes
that correspond to a socket.