feat(build: qemu): add network device to qemu CLI arguments
For easy debugging we use a tap interface as our backend. This lets
us use tcpdump on the tap interface to see what's going on live.
We also dump all network traffic into the build/network.dat file, so
that it can be analyzed with wireshark.

This argument is only added if there is a tap interface currently
up on the system. The interface name is defined by the QEMU_TAP_IF
enveironment variable (default: tap0).