feat(net: packet): add packet_set_l2/l3_size()
These functions are used to eaily configure the start of an already
build packet. This is typically the case with the RX path, where we
build an entire packet from the data received from the NIC.

This commit also adds the concept of 'payload' to refer to the packet's
content beyond layer 3, and 'header' for L2/L3 headers.