feat(spinlock): add locked_scope macro
This is particularily useful when taking and releasing a lock
in a sequential manner. For example, to modify a list that is
protected by a spinlock, you may do:

locked_scope(&list_lock) {
    llist_add(&global_list, &new_item);
}

In this example the @c list_lock spinlock will automatically be taken when
entering the scope, and automatically released when leaving.

WARNING: As this macro uses a for loop to function, any 'break' directive
placed inside it will break out of the guarded scope instead of that of its
containing loop.