feat(sched): add no_scheduler_scope() macro
Similarily to the locked_scope() macro, this one can be really useful
when needing to when we need to sequentially lock() and unlock() the
scheduler.

This lets us create scopes during which we want to be sure that
we will not rescheduled, and has the advantage of making it absolutely
clear when reading the code.

Example:

no_scheduler_scope () {
    // scheduler sensitive code ...
}