feat(misc: worker): use a single thread for all executions
We were creating a new thread each time we needed to run the worker.
This is no longer the case. Instead, we added a required worker_init()
function that allocates a thread, and we simply sched_block/unblock_thread()
it when entering and exiting the wrapped function.

This has the disavantadge of having to now explicitely initialize the
worker struct, but it saves us a lot of PIDs and execution time.