wip: feat(net: ethernet): handle received packets inside a worker thread
The old way of doing things (doing everything when receiving the packet)
had the huge disadvantage of performing possibly time expensive actions
in a un-interruptible context. In other words, we were blocking the
whole kernel everytime we received a new packet.

Now, everything is done by a separate worker thread, outside of the
interrupt handler. We have to bufferize all the packets that come until
the worker thread is scheduled.