feat(net: ipv4): raw socket receive a cloned version of the packet
Raw packets may receive packets that will also be sent over to regular
sockets. This means that the packets are succeptible to be added to
multiple RX queues at a time, which we don't want.

Since only one regular sockets should ever receive the packet (multiple
sockets cannot bind a same address), it is simpler to send the original
packet to its 'true' destination and receive a 'copy' of the packet
inside raw sockets.