fix(i686: process): unlock scheduler inside arch_thread_entrypoint()
The scheduler lock is usually taken by the previous thread an released
by the newly scheduled one inside the schedule() function, but this does
not apply to newly created threads. Since they were not previously
re-scheduled by the schedule() function, we must absolutely imitate this
if we don't want to be deadlock when scheduling the next thread.