refacto(sys: process): allocate kernel stack in arch independent codepath
This should always be the same regardless of the architecture. The only
reason not to have done so earlier was because the kstack top address is
located inside the arch-dependent process context structure.

We now use generic thread_kernel/user_stack_get/set() functions
to modify/read the thread's stack addresses. This makes it both
generic and more readable than the previous version.

We also switched over to using the raw VM API for allocting the stack
since it is faster, and that it makes it guaranteed that the stack
fits known rules (page aligned).