refacto(libalgo: linked_list): use a list head node
Our previous design did not include a list head node. To perform
boudary checks such as llist_is_empty() or checking if there is
a next/prev node, we had to compare against a potential NULL value.

This has multiple disadvantages:
1. Prone to errors that would cause a NULL-reference exception
2. We had to keep track of the list's head when inserting and removing
3. Slower

This commit changes things for a new design that simply adds a list head
node. When the list is empty, the head points to itelf (next/prev).
This means that the list is never completely empty, and that we can
now easily update the prev->next and next->prev pointers when inserting,
or removing an item from the list.

Addition and deletion can now safely be done in O(1) time complexity.

API changes:
- llist_t is now a typedef to the struct linked_list_head
- queue_t is now a typedef to llist_t (structure is equivalent)
- llist_remove() does not take in the list's head anymore
- add llist_add_after/before() to insert an element at an arbitrary
  position inside the list.

Signed-off-by: Léo DUBOIN <leo@duboin.com>