feat(sys: process): add thread_fork() function
This function essentially performs what's described in man 2 fork.

It creates a new process, and initialize its initial thread to be a
1:1 copy of the forked thread.

With what we've implemented so far, this means:
* Duplicating the original process's information
* Duplicating the original process's address space
* Duplicating the original thread's execution state

As this is meant to be the only way of creating new processes, the
new thread_fork() function has replaced the previously unused
process_create().

TODO:
* Copying registers from interrupt frame before starting the thread