feat(toolchain: newlib): verify automake/autoconf version
When trying to build newlib using a recent automake (1.17.1), the name
of the output libm object files does not match the one hardcoded inside
the Makefile.am file anymore.

Since newlib is known to compile with automake 1.15 and autoconf 2.69,
we now make sure that we are using these versions before building.