refacto(toolchain): generate a proper sysroot
We previously used a dummy sysroot made up of minimal libc headers
(stdio, ...). Now that we are able to compile a full libc, and have
a uapi include folder, we can generate an actual sysroot with all our
build artifacts.

Our custom toolchain is now placed inside our sysroot, instead of the
opt folder used previously. This should maybe not be done like so,
since it would mean that our toolchain is runnable from our userland
(which is far from being the case), but at least everything is kept
together somewhere.

Signed-off-by: Léo DUBOIN <leo@duboin.com>