refacto(fs: vfs): rename vfs_*_at() function to vfs_*()
In the future (aka the very next commit), we will want to perform these
actions (create, remove, ...) directly on a vnode directly, without
having to parse the entire path again.

The *_at() version of the functions have been replaced with the version
that takes in a vnode instead of a path.

The E_NOT_IMPLEMENTED error value returned in case the vnode does not
expose the necessary operators has been changed to E_NOT_SUPPORTED.