feat(devices: uart): do not update file position when reading/writing
The concept of position inside the file does not apply to the UART
device. It is only possible to send data through this kind of device,
without any kind of offset mechanism.

We removed the reference to the backing pseudo file's offset to avoid
any confusion. Plus, it should save us unnecessary clock cycles inside
a hot path.