feat(i686: timer): do not try to compensate for invalid frequencies
We now rely heavily on the HZ constant being exact for conversions between
seconds and clock ticks. Moreover, this design choice was just plain
wrong to begin with. We were 'hiding' the error under the rug, instead
of informing the caller and let them correct their own mistake. We were
trying to be too smart.