feat(fs: vfs): add vnode statistics
As we haven't implemented user permissions yet, the mode stat is not
meaningful and can only be parsed from the on-disk metadata (only
applicable for tar filesystem).

Same for the device number (both types) and block size. We haven't
implemented those API, so these fields are left empty until possible.

For the modified/accessed/changed times, we do not have a system-wide
real time clock mechanism and use the system's timer as a placeholder
for now.