refacto(sys: interrupts): split arch-specific setup from interrupt handling
The actual generic interrupt routines and custom handlers have nothing
to do with the architecture. The only thing that could cause a problem
here is that we still rely on an arch-dependant interrupt frame structure.
But, in practice, the actual interrupt handlers should be arch-specific,
so this may not be a real problem in the end ... we'll see ...