feat(memory: vm): add VM_FIXED allocation flag
This flags is meant to be used in combination with a starting address
hint. When used, if no free memory area is availabe at the exact location
specified by the argument, the allocation fails.

This is usefull when dealing with strict memory placement requirements,
for example when loading an executable files whose content MUST be
placed at the specified address.