refacto(memory: as): stop inserting/removing memory segment from VMM code
The insertion and deletion of the VM segment is independent from the VMM
code. It has now been moved up the callstack, and is now performed by
the vm_allocate()/vm_free() functions.

This change also lets us remove the vm_segment_insert/remove() functions
from the header, that I was not a fond of having shared with every
compilation unit that imported it in the first place.