feat(memory: as): add vm_segment resizing
It is now possible to resize an already existing virtual memory segment.

When increasing its size, the newly mapped memory follows the same
principles as a newly created segment, being mapped to a physical page
only during a page fault. When reducing its size, the memory that was
previously part of the segment gets free'd.