fix(sys: sched): NULL reference in schedule_locked() when queue is empty
We were using queue_peek() == NULL to detect the presence of any other
threads inside the ready queue. This worked originally because the first
queue implementation returned a NULL value in this case, but this is no
longer the case since switching to the sentinel version.

This commit replaces this check with queue_is_empty() instead, which
also makes it clearer what we are actually checking for.