fix(i686: mmu): panic when address_space_fault() fails during page fault
We did not check whether the fault handler actually succeeded, or
detected an invalid fault. This caused an infinite and silent chain
of page faults to be raised on the same address (since the faulty
instruction gets re-executed after a fault) when an actual page
fault occurs (e.g NULL pointer reference).