fix(sys: process): do not initialize the address space when forking
There is no need to initialize the address space when forking, since
we are copying the content of the original one over.

All that is done when iniitializing an address space is to setup
the VMM. But, all userland addresses are duplicated using CoW
(including all VMAs), and the content of the VMM is also copied
anyway.

We had to move the locking and list initialization out of the
address_space_init() function since those are not touched during
the copy operation.