fix(sys: sched): prevent the current thread from killing itself
In the case where there is only one running thread, the scheduler
skips the IDLE thread and instantly reschedules it. This causes a
problem when the re-scheduling is caused by the current thread calling
thread_kill() on itself.

Since the thread_free() function is delayed until the killed thread
gets rescheduled, this situation causes the thread_free() function
to be executed on the currently running thread. This obviously
causes the kernel to crash when eventually freeing the thread's kernel
stack, that is still in use in this case.

This commit prevents the scheduler from rescheduling the current thread
if it is in the KILLED state, and adds a safety check inside the
thread_free() to make this error obvious if it ever happens again.