fix(memory: as): dynamically allocate list heads
For the segment and kmalloc lists to be identical between a forked
process and its parents, the address of the sentinels must absolutely
be kept the same. If not, the whole "sentinel detection" mechanism
required for the doubly linked lists to work gets screwed.

This was the case with the previous design, with list sentinels embeded
inside the address space struct, which gets re-allocated and copied into
during a fork.

The address_space_clear() function now also releases the memory
allocated for these heads, since it must be done while the address space
is still loaded. This implies that all calls to address_space_clear()
expecting the address space to be reused afterwards MUST now call
address_space_init() in order to re-allocate those.