refacto(sys: process): clear process's address space before killing
The process must still be loaded for its address to be cleared. This
can only be done before marking it as killed since it will be free'd
by another process's thread when releasing its last remaining reference.

This change also allows us to release the thread's user space, that
could not be free'd because of the exact same reasons.