feat(device: console): use kernel console as the default stdout
This commit adds a 'console' device and mounts the devtmpfs onto
the /dev folder during startup. The resulting '/dev/console' is
used as the default standard output (and input even though it
currently does not support reading from), and installed inside
the kernel process' open file descriptions table so that it can
then be inherited by the first process.