feat(misc: execfmt): pass arguments to the executable
These arguments are passed to the executable's entrypoint function.

As per usual we follow the system-v ABI, meaning that the arguments
are pushed onto the stack, and the starting stack pointer is set to
point to the argc parameter. Note that we also had to push an empty
stack base pointer along with the arguments, to emulate a regular
function call.

The init process does not take any arguments, thus the content
of its argv parameter only contains the executable's path (argv[0]).