sys/process: stop using current->context.esp in arch_thread_entrypoint()
This commit changes the way we keep track of the starting stack pointer
for a newly forked thread.

As specified inside the new comment, the thread context's esp value
is updated everytime an interrupt is triggered. This includes pagefault
interrputs that are very likely to occur when initializing the new
process's address space. If one were to be triggered before creating
the new thread, the latter would start with a user stack located inside
the kernel addres space.

To avoid future problems, we do not use the thread context's ESP value
for this purpose anymore, and instead push the stack pointer onto
the entrypoint stack inside arch_thread_spawn().

Signed-off-by: Léo DUBOIN <leo@duboin.com>