feat(main): mount initfs using an init ramdisk
This commit finally gets rid of the TODO 'replace with a ramdisk'
the kernel_mount_initfs() function.

As previously specified, we assume that the multiboot structure
always contains a single module at most, and that it is guaranteed
to be the initramfs. Since this is the case, we do not need to relocate
the module's content anymore.

Signed-off-by: Léo DUBOIN <leo@duboin.com>