devices/block: add page cache
This commit adds a simple page cache mechanism for block devices.
It is designed to be able to contain multiple blocks per page so it
can be used as both a page cache and a buffer cache.

Each block device comes with its own page cache, from which the
buffer/pages are looked for by the higher layers (vnode, filesystem).
All cached pages are stored in a global list, that is then used by
a kernel thread to regularly flush the content of modified pages to
disk.

Cache entries are obviously reference counted since they will be shared
by multiple users. The reference count is incremented when retreiving
the entry in block_device_cache_get(), and decremented when releasing
it in block_device_cache_put().

Future possible improvements:
- Cache page_cache_entry structures (slab allocator)
- Implement a real scheduling algorithm in the writeback thread
- Add 'linked' entries (used to share pages between page & buffer cache)

Signed-off-by: Léo DUBOIN <leo@duboin.com>