fs: socket: allocate vnode in kernel space
The socket's vnode structure was allocated using the KMALLOC_DEFAULT
flag. There is no real reason for it to be allocated this way, and
contrary to this this would only create potential security issues.

Moreover, we are trying to get rid of the KMALLOC_DEFAULT & KMALLOC_KERNEL
flags, and only allow the kmalloc function to allocate kernel memory.

Use the KMALLOC_KERNEL flag when allocating vnodes for sockets.