memory: kmalloc: remove KMALLOC_DEFAULT flag
The kmalloc function should only be used to allocate buffers used by
the kernel. The only user memory that should be allocated in userland
by the kernel is the user executable and user stack, which should be
page aligned and allocatable through the VM api.

Remove the KMALLOC_DEFAULT flag.