memory: check for NULL address space in page_fault handler
The current process's is NULL until we initialize the kernel process
in process_init_kernel_process(). Things get messy and memory is
corrupted if a NULL reference exception happens while in this state.

When the page fault exception is triggred by a valid lazy allocation,
the address space used should be &kernel_address_space and not that
of the currently running process.

Make sure that an address space is defined, and panic if not the case.