fs/tar: fix invalid address in tar_node_free()
We were passing the address of the tree node to the kfree() function
instead of the tar node. This worked fine until now since the former
is embedded inside the latter, but does not work anymore since the
switch to a slab allocator.

Freeing the correct address instead fixes the following error:

kmem_cache_free:419: failed to find bufctl for c09f6700 (align: 16, obj: c09f6700)
Call stack:
  #0  0xc010e337 in <kmem_cache_free+330>
  #1  0xc011173e in <kfree+136>
  #2  0xc010342d in <tar_node_free+57>
  #3  0xc010348e in <tar_node_from_header+92>
  #4  0xc01037c5 in <tar_init_tree+471>
  #5  0xc0103dbb in <tar_new+408>
  #6  0xc0101fc7 in <vfs_mount_at+83>
  #7  0xc010204c in <vfs_mount_root+57>
  #8  0xc010120d in <kernel_mount_initfs+130>
  #9  0xc0101477 in <kernel_main+530>
  #10  0xc011571b in <higher_half+31>
===