sys/process: do not free the startup thread's structure
The kernel_process_initial_thread (TID 0) is statically allocated
and not kmalloc'd like all other threads in the kernel. This means
that we cannot call kfree() on it as we usually would.

We should be able to release this memory as well as other structures
that are only required during the setup process, but this is another
matter.

Do not free the initial kernel thread in thread_free(). This fixes
the following error:

[process] terminating thread 0 (kstartup)
[vm] freeing unaligned virtual address: c0954990 (skipping)
kfree:49: !(page->flags & PAGE_SLAB)
Call stack:
  #0  0xc0111693 in <kfree+104>
  #1  0xc0105c86 in <thread_free+186>
  #2  0xc0105ccb in <thread_switch+28>
  #3  0xc0104a07 in <schedule_locked+280>
  #4  0xc0104a4d in <schedule+34>
  #5  0xc0118925 in <irq_timer_handler+111>
  #6  0xc0115de7 in <default_interrupt_handler+387>
  #7  0xc0115ec7 in <__common_handler+7>
  #8  0xc01032cb in <tar_read+146>
  #9  0xc0103bff in <tar_file_read+195>
  #10  0xc010c0ba in <map_file+248>
  #11  0xc0107e83 in <execfmt_execute+148>
  #12  0xc0105ed0 in <__process_execute_in_userland+65>
  #13  0xc01172b1 in <arch_thread_entrypoint+233>
===