memory/vm: add cache policy configuration flags
The caching policy used for a vm_segment is now explicitely configured
inside the segment's flag. The chosen policy is forwarded to the MMU API
when mapping an address inside the segment.

This commit also adds the vm_set_policy() function that can be called
to dynamically re-configure these policies for an already existing
segment.

When no policies are specified the API defaults back to using write-back
caching (regular caching).