i686: process: clear eax register before child returns from fork
EAX is supposed to contain the return value after a syscall. Because the
fork() syscall does not return to userland the same way other syscalls
do, we never end up setting this return value.

Always set the content of the EAX register to 0 before jumping to userland
to fix this issue. This is the value returned by the child process,
which we always are in this codepath, and is a sane default for future
proofing.