i686: mmu: flush recursive page table address during COW
The PD entry is modified when duplicating a CoW pagetable, meaning
that we must flush the recursive pagetable address before accessing
and modifying the pagetable's content. This also stands when setting
up the CoW mappings.

This fixes a stack corruption issue where the pages would not be
duplicated when accessed by the parent after a fork, causing the
child to start executing inside the parent's codepath (skipping
the common pid==0 check performed on the returned pid value).

This also fixes the following CoW error that often appeared after a fork:

[process] forked new process: init (5)
[mmu] {5} cow @ 110000
[mmu] page->refcount = 2 @ c015a828
[mmu] page->refcount = 1 @ c015a828
[mmu] {1} cow @ 110000
[mmu] page->refcount = 2 @ c015a870
[mmu] regular ro page (110f9c)
[vm]  {1} cow @ 110f9c failed: Operation not permitted
[mmu] {1} cow @ 800c000
[mmu] page->refcount = 2 @ c015a830
[mmu] page->refcount = 1 @ c015a830
[mmu] {5} cow @ 800c000
[mmu] page->refcount = 2 @ c015a888
[mmu] regular ro page (800c008)
[vm]  {5} cow @ 800c008 failed: Operation not permitted